[REF] Simply package/module structure naming
