+++ esg_matching/__init__.py
@@ -1,8 +1,8 @@
 """ The esg-matching library provides the following packages:
 
-    - **db_data_source:** provides representations of data sources connected to a database table
-    - **db_engine:** provides database connectivity and processing (for queries and data manipulation)
-    - **db_matcher:** main package to perform exact matching based on database SQL statements.
+    - **data_source:** provides representations of data sources connected to a database table
+    - **engine:** provides database connectivity and processing (for queries and data manipulation)
+    - **matcher:** main package to perform exact matching based on database SQL statements.
     - **exceptions:** centralized exception management for the esg-matching library.
     - **file_reader:** file processing capabilities.
