+++ esg_matching/file_reader/__init__.py
@@ -1,9 +1,9 @@
 """ The **esg_matching.file_reader** package provides the following modules for file processing.
 
     - **file.py:** used to represent a file object.
-    - **file_reader.py:** abstract representation of a file-reader object.
-    - **file_reader_csv.py:** concrete implementation of a file-reader object for csv files.
-    - **file_settings.py:** used to represent settings based on json files to perform database-driven matching.
-    - **file_utils.py:** contains utility functions associated to file management.
+    - **base_reader.py:** abstract representation of a file-reader object.
+    - **csv_reader.py:** concrete implementation of a file-reader object for csv files.
+    - **settings.py:** used to represent settings based on json files that describe matching rules.
+    - **utils.py:** contains utility functions associated to file processing.
 
 """
