+++ esg_matching/file_reader/file.py
@@ -1,7 +1,10 @@
-""" Base class allows to create an object to represent a file accompanied with its matching settings. """
+"""
+    The **esg_matching.file_reader.file** module provides a representation of a file object
+    independently of the application environment.
+"""
 
-from esg_matching.file_reader.file_settings import JsonSettings
-from esg_matching.file_reader import file_utils
+from esg_matching.file_reader.settings import JsonSettings
+from esg_matching.file_reader import utils
 from esg_matching.exceptions import exceptions_file
 
 
@@ -16,7 +19,7 @@ class File:
                 File extension (e.g. .csv)
             _filename_settings (str)
                 Path and name of the json file used as the matching settings.
-            _obj_settings (esg_matching.file_reader.file_settings.JsonSettings)
+            _obj_settings (esg_matching.file_reader.settings.JsonSettings)
                 Settings object that translates the properties of a json file into matching parameters and policies.
     """
 
@@ -54,18 +57,24 @@ class File:
 
     def _get_file_info(self):
         """
-        Private method to retrieve the complete path, name and extension of the file described in the json settings.
+        Private method that retrieves information about the file described by its json settings.
+        This method checks the file_path in the json file to verify if a file or a folder was defined.
+        In the case of a folder, the application looks for the newest file in that folder that follows the
+        file_extension and filename_pattern informed in the json file. In the case of a file, the application
+        checks if the file exists. As a result, for both cases, it updates the following internal
+        attribute variables: self._filename and self._file_extension.
+
         """
 
-        if file_utils.is_file(self._obj_settings.file_path):
+        if utils.is_file(self._obj_settings.file_path):
             self._filename = self._obj_settings.file_path
-            self._file_extension = file_utils.get_extension(self._filename)
+            self._file_extension = utils.get_extension(self._filename)
         else:
-            if not file_utils.is_dir(self._obj_settings.file_path):
+            if not utils.is_dir(self._obj_settings.file_path):
                 raise exceptions_file.FilePathNotFoundOrUnreachable
-            self._filename = file_utils.get_newest_filename(self._obj_settings.file_path,
-                                                            self._obj_settings.file_extension_pattern,
-                                                            self._obj_settings.filename_pattern)
+            self._filename = utils.get_newest_filename(self._obj_settings.file_path,
+                                                       self._obj_settings.file_extension_pattern,
+                                                       self._obj_settings.filename_pattern)
             if self._filename == '':
                 raise exceptions_file.FileNotFoundOrUnreachable
-            self._file_extension = file_utils.get_extension(self._filename)
+            self._file_extension = utils.get_extension(self._filename)
