+++ esg_matching/matcher/dfm.py
@@ -1,8 +1,8 @@
 """ Base class allows to create a matcher that performs direct full matching on a database """
 
-from esg_matching.db_matcher.db_matcher import DbMatcher
-from esg_matching.db_engine.engines.connector import DbConnector
-from esg_matching.db_matcher.matching_policy import MatchingPolicy
+from esg_matching.matcher.db_matcher import DbMatcher
+from esg_matching.engine.connectors.base_connector import DbConnector
+from esg_matching.matcher.policy import MatchingPolicy
 from esg_matching.exceptions import exceptions_matching_policy
 
 
@@ -21,7 +21,7 @@ class DbMatcherDfm(DbMatcher):
             Constructor method.
 
             Parameters:
-                db_connector (DbConnector): database engines
+                db_connector (DbConnector): database connectors
 
             Returns:
                 DbMatcherDfm (object)
