+++ pyproject.toml
@@ -50,7 +50,6 @@ pandas = ">=1.1.5"
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.5"
 black = "^21.8b0"
-pre-commit = "^2.15.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
